📑 :: Change registration page type and code cleanup
