+++ app/src/main/res/layout/activity_main.xml
@@ -49,7 +49,7 @@
             android:layout_width="match_parent"
             android:layout_height="0.5dp"
             android:orientation="horizontal"
-            app:layout_constraintGuide_percent="0.92"/>
+            app:layout_constraintGuide_percent="0.92" />
 
         <androidx.constraintlayout.widget.ConstraintLayout
             android:id="@+id/main_app_bar_layout"
@@ -75,7 +75,7 @@
                 android:layout_height="wrap_content"
                 android:background="@drawable/ic_profile"
                 app:layout_constraintEnd_toEndOf="parent"
-                app:layout_constraintTop_toTopOf="parent"/>
+                app:layout_constraintTop_toTopOf="parent" />
 
             <ImageView
                 android:id="@+id/alarm_btn"
@@ -101,9 +101,9 @@
             android:name="androidx.navigation.fragment.NavHostFragment"
             android:layout_width="match_parent"
             android:layout_height="0dp"
-            app:layout_constraintTop_toTopOf="@id/guideline4"
-            app:layout_constraintBottom_toBottomOf="@id/guideline5"
             app:defaultNavHost="true"
+            app:layout_constraintBottom_toBottomOf="@id/guideline5"
+            app:layout_constraintTop_toTopOf="@id/guideline4"
             app:navGraph="@navigation/bottom_nav_graph" />
 
         <com.google.android.material.bottomnavigation.BottomNavigationView
@@ -112,11 +112,11 @@
             android:layout_height="0dp"
             android:background="@color/light_black"
             app:itemBackground="@drawable/selector_nav_bg"
-            app:itemIconTint="@drawable/selector_nav_color"
             app:itemIconSize="60dp"
+            app:itemIconTint="@drawable/selector_nav_color"
             app:labelVisibilityMode="unlabeled"
-            app:layout_constraintTop_toBottomOf="@id/guideline5"
             app:layout_constraintBottom_toBottomOf="parent"
+            app:layout_constraintTop_toBottomOf="@id/guideline5"
             app:menu="@menu/bottom_navigation_menu" />
